feat: hook datadog-agent collect log (#9)

+++ charts/spartan/templates/_hook.tpl
@@ -37,7 +37,7 @@ spec:
       securityContext:
           {{- toYaml .Values.podSecurityContext | nindent 8 }}
       restartPolicy: {{ .hook.restartPolicy | default "Never" }}
-      {{- if .Values.datadog.enabled }}
+      {{- if and (.Values.datadog.enabled) (.hook.collectLog) }}
       shareProcessNamespace: true
       {{- end }}
       containers:
@@ -54,7 +54,7 @@ spec:
             - {{ default "/bin/sh" .hook.shell }}
             - -c
             - |
-            {{- if .Values.datadog.enabled }}
+            {{- if and (.Values.datadog.enabled) (.hook.collectLog) }}
               trap 'sleep 10 && pkill agent' EXIT
               set -o pipefail
               if [ ! `which curl` ]; then sleep 300; else while ! curl -Ns localhost:8126; do sleep 1 && echo "Waiting for datadog agent to start...."; done; fi
@@ -82,7 +82,7 @@ spec:
                 name: {{ .Values.configMap.externalConfigMapEnv.name }}
               {{- end }}
           env:
-          {{- if .Values.datadog.enabled }}
+          {{- if and (.Values.datadog.enabled) (.hook.collectLog) }}
             - name: DD_KUBERNETES_KUBELET_NODENAME
               valueFrom:
                 fieldRef:
@@ -130,12 +130,14 @@ spec:
               mountPath: {{ .sharedVolume.mountPath }}
               subPath: {{ .name }}
             {{- end }}
-            {{- end }}
-
+          {{- end }}
+        {{- $hook := .hook }}
         {{- range $sidecar := .Values.sidecars }}
-        {{ include "sidecar.template" (dict "sidecar" $sidecar "Values" $.Values "Chart" $.Chart "Release" $.Release) | indent 8 }}
+          {{- if and (eq $sidecar.name "datadog-agent") ($hook.collectLog) }}
+            {{ include "sidecar.template" (dict "sidecar" $sidecar "Values" $.Values "Chart" $.Chart "Release" $.Release) | indent 8 }}
+          {{- end }}
         {{- end }}
-        {{- with .Values.nodeSelector }}
+      {{- with .Values.nodeSelector }}
       nodeSelector:
           {{- toYaml . | nindent 8 }}
         {{- end }}
